requirements.txt
geographiclib==1.50
geopy==1.20.0
Pillow==7.0.0
python-dotenv==0.10.3